Update 1.4
Added a few more PokeBalls. Ran into some issues with developer cheat codes and user input, completely reworked the system but the same issue remains. Will work more later.

+++ src/PokeBall.java
@@ -1,11 +1,11 @@
-public class PokeBall extends CatchingBall{
+public class PokeBall extends CatchingBall {
 
     public PokeBall() {
-        this.setBaseCatchRate(1);
+        this.setBaseCatchRate(1.0);
         this.setDisplay("Poke Ball");
         this.setId("poke-ball");
     }
-    public int getCatchMultiplier(Pokemon opponent) {
+    public double getCatchMultiplier(Pokemon opponent) {
         return this.getBaseCatchRate();
     }
 }
